Release v1.6.0 (#111)
* redeemCoupon docs

* Test Order.findByPaymentId

* Test Order.findByPaymentId null

* Order status check fix

* Order findByPaymentIdOrFail

* Simplify Order.findByPaimentId

* guardMollieMandate

* helper money_to_decimal

* guardMollieMandate

* restartCycle

* improved Mandate retrieval and validation

* Use temporary payment_id to fix race condition

* SubscriptionBuilders skipTrial

* Improved mandate retrieval - status 410

* Preparing release v1.5.0

* Prefix temporary uuid

* Actually test findByPaymentIdOrFail

* Handle removing inexistent scheduled items

* MandateClearedFromBillable fix #109

* Use SerializesModels trait

* Prepare release v1.6.0

+++ src/Subscription.php
@@ -282,7 +282,7 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     {
         $item = $this->scheduledOrderItem;
 
-        if($item->isProcessed(false)) {
+        if($item && $item->isProcessed(false)) {
             $item->delete();
         }
